+++ memo/ChangeLog.md
@@ -136,6 +136,7 @@
   - main (`ble/bin/.freeze-utility-path`): fix unupdated temporary implementations `#D1528` c70a3b4
 - test (`test-canvas`): fix dependency on `ext/contra` `#D1525` c89aa23
 - util: inherit special file descriptors `#D1552` 98835b5
+- global: use `_ble_term_IFS` 0000000
 
 <!---------------------------------------------------------------------------->
 # ble-0.4.0-devel2